fix(step-ca): step-ca configuration
step-ca was a mess to configure, because the URLs provided at the
directory URL API are constructed from the "address" and "port" fields
in this config, it makes quite harder to have this behind a reverse
proxy. This fix relies on the tailscale IP of the machine which all
other server should have access to, and correctly adding this IP to the
certificate through the "dnsNames" entry.

Therefore, other machines can now correctly access the directory URL
through https://ca.mesh.gq/acme/acme/directory and also the different
endpoints using https://100.64.0.5:6060/[...].