fix(nodes): certificate renewal
Since the SNI proxy implementation, TLS trafic is redirected base on
SNI. However, to renew TLS certificates, we are using the HTTP-01
challenge. The HTTP trafic was not redirected correctly. This patch
introduce VHOST entries to strix, and redirect HTTP traffic to the
correct machine for ACME challenge validation. On the proxied machine,
we are listening on IPV6 address port 80 (we could be more restrictive).